Improved variables + Animations 95% done
Last commit of the night.
Fixed the variables.
Partially fixed tab animations (Double tab animations needs to be fixed,
and by double I mean both an animation ontop and one on the bottom)

+++ src/main/java/io/puharesource/mc/titlemanager/TitleManager.java
@@ -128,7 +128,7 @@ public class TitleManager {
     }
 
     public static void removeRunningAnimationId(int id) {
-        runningAnimations.remove(id);
+        runningAnimations.remove((Integer) id);
     }
 
     public static List<Integer> getRunningAnimations() {
